+++ app/controllers/passwords_controller.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+# Override for devise passwords controller.
+class PasswordsController < Devise::PasswordsController
+  layout "full_page"
+end