+++ test/controllers/internal_controller_test.rb
@@ -4,20 +4,24 @@ require "test_helper"
 
 # Tests to assure internal pages are accessible for logged in users.
 class InternalControllerTest < ActionDispatch::IntegrationTest
+  setup do
+    @regular_user = users(:one)
+  end
+
   test "should get dashboard" do
-    # login(@regular_user)
+    login(@regular_user)
     get dashboard_url
     assert_response :success
   end
 
   test "should get survey" do
-    # login(@regular_user)
+    login(@regular_user)
     get survey_url
     assert_response :success
   end
 
   test "should get thank you" do
-    # login(@regular_user)
+    login(@regular_user)
     get thank_you_url
     assert_response :success
   end