+++ CHANGELOG.md
@@ -3,4 +3,9 @@
 ### Enhancements
 - **General Changes**
   - Setup initial flow through website
+  - New users can now create accounts, confirm their new accounts, and reset
+    their passwords
+  - New users are sent a welcome email
+  - New users are sent an email verification email to confirm their email
+    address
   - Initial Ruby on Rails 5.1.2 framework